@@ -9,57 +9,42 @@
   "dist": "dist",
   "mainInput": "main",
   "mainOutput": "main",
+  "scripts": {
+    "start": "babel-node server.js"
+  },
   "dependencies": {
-    "babel-runtime": "^5.8.24",
-    "bootstrap-sass": "^3.3.5",
+    "babel-runtime": "^6.2.0",
     "classnames": "^2.1.3",
     "color": "^0.10.1",
-    "extract-text-webpack-plugin": "^0.8.2",
-    "lodash": "^3.10.1",
-    "react": "0.13.x",
-    "react-bootstrap": "^0.25.2",
-    "react-google-maps": "^3.0.0",
-    "react-redux": "^3.0.1",
-    "react-router": "^1.0.0-rc1",
+    "extract-text-webpack-plugin": "^0.9.1",
+    "history": "^1.13.1",
+    "react": "0.14.x",
+    "react-dom": "^0.14.3",
+    "react-google-maps": "^4.6.1",
+    "react-redux": "^4.0.0",
+    "react-router": "^1.0.0",
     "redux": "^3.0.2",
     "redux-logger": "^2.0.0",
     "redux-thunk": "^1.0.0",
-    "whatwg-fetch": "^0.9.0"
+    "whatwg-fetch": "^0.10.1"
   },
   "devDependencies": {
-    "autoprefixer-loader": "^3.1.0",
-    "babel": "^5.8.21",
-    "babel-eslint": "^4.1.3",
-    "babel-loader": "^5.3.2",
+    "babel-cli": "^6.2.0",
+    "babel-core": "^6.2.1",
+    "babel-eslint": "^5.0.0-beta4",
+    "babel-loader": "^6.2.0",
     "babel-plugin-react-transform": "^1.1.1",
-    "css-loader": "^0.19.0",
+    "babel-preset-es2015": "^6.1.18",
+    "babel-preset-react": "^6.1.18",
+    "babel-preset-stage-0": "^6.1.18",
     "eslint": "^1.2.1",
     "eslint-loader": "^1.0.0",
     "eslint-plugin-react": "^3.2.3",
     "exports-loader": "^0.6.2",
-    "grunt": "^0.4.5",
-    "grunt-contrib-clean": "~0.6.0",
-    "grunt-contrib-connect": "~0.11.2",
-    "grunt-contrib-copy": "~0.8.1",
-    "grunt-karma": "~0.12.1",
-    "grunt-open": "~0.2.3",
-    "grunt-webpack": "^1.0.11",
     "imports-loader": "^0.6.4",
-    "jasmine-core": "^2.3.4",
     "json-loader": "^0.5.3",
-    "karma": "~0.13.10",
-    "karma-jasmine": "^0.3.5",
-    "karma-phantomjs-launcher": "~0.2.1",
-    "karma-script-launcher": "~0.1.0",
-    "karma-webpack": "^1.5.0",
-    "load-grunt-tasks": "~3.3.0",
-    "node-sass": "^3.3.3",
-    "react-transform-catch-errors": "^1.0.0",
-    "react-transform-hmr": "^1.0.1",
-    "redbox-react": "^1.1.1",
+    "react-hot-loader": "^1.3.0",
     "redux-devtools": "^2.1.5",
-    "sass-loader": "^3.0.0",
-    "style-loader": "^0.12.3",
     "url-loader": "^0.5.6",
     "webpack": "~1.12.2",
     "webpack-dev-server": "^1.10.1"
