@@ -1,13 +1,12 @@
 'use strict';
 
-import 'babel-core/polyfill';
 import React from 'react';
+import { render } from 'react-dom';
 import Root from './containers/Root';
-import './styles/main.scss';
 
 const content = document.getElementById('content');
 
-React.render(
+render(
   <Root />,
   content
 );
