@@ -1,4 +1,5 @@
-import { combineReducers } from 'redux';
+//import { combineReducers } from 'redux';
 
-export default combineReducers({
-});
+export default function(state) {
+    return state;
+};
